requirements.txt
asyncio==3.4.3
influxdb==5.3.1
requests==2.26.0
urllib3==1.26.7
python-decouple==3.6



